Add files via upload
This report analyses the structure of Linux. Explains how system ageing works on computer devices. We explored the necessary hardware components that are required for the Linux system. Explain the difference between lightweight and heavyweight OS. The report also highlights the benefits and strengths of each design and contains the author's experience and recommendations in the report.